+++ CHANGES.md
@@ -31,7 +31,9 @@ The ['change-log' (below)](#changelog) in this file should summarize **all API c
     - Basic 'count' functionality to report the number of files and tasks in each catalog.
 - Added subcommands for git repositories in [astrocats/catalog/catalog.py](https://github.com/astrocatalogs/astrocats/blob/master/astrocats/catalog/catalog.py).
     - The 'push' subcommand can now be used to add, commit and push all data files in each data repository.  This works in all installed catalogs.
-
+- `astrocats/catalog/entry.py`
+    - `Entry.add_alias` [new-function]
+        - New method to add aliases to an existing entry after first 'cleaning' the alias name - in the same way as the entry names are cleaned by the containing catalog.  In this way, the stored aliases should be guaranteed (in general) to match the corresponding entry names (and naming styles).
 
 <a name='v0.2.0'>
 ### v0.2.0 - 2016/07/18 ###